Fix race when calculating db folder size.
